+++ packages/components/src/Form/Inputs/Select/Select.tsx
@@ -24,17 +24,21 @@
 
  */
 
-import React, { forwardRef, Ref, FormEvent } from 'react'
+import React, { forwardRef, Ref, FormEvent, ReactNode } from 'react'
 import styled from 'styled-components'
 import { CustomizableAttributes } from '@looker/design-tokens'
+import { Box } from '../../../Layout'
 import { ListItem } from '../../../List'
+import { Heading, Paragraph } from '../../../Text'
 import { ValidationType } from '../../ValidationMessage'
 import {
   Combobox,
   ComboboxInput,
   ComboboxList,
   ComboboxOption,
+  comboboxOptionGrid,
   ComboboxOptionObject,
+  ComboboxOptionText,
   ComboboxProps,
   getComboboxText,
 } from '../Combobox'
@@ -47,18 +51,16 @@ export const CustomizableSelectAttributes: CustomizableAttributes = {
   py: 'none',
 }
 
-export type OptionsType<OptionType> = OptionType[]
-
-export interface GroupType<OptionType> {
-  options: OptionsType<OptionType>
-  [key: string]: any
+export interface SelectOptionObject extends ComboboxOptionObject {
+  description?: string | ReactNode
 }
 
-export type GroupedOptionsType<UnionOptionType> = Array<
-  GroupType<UnionOptionType>
->
+export interface SelectOptionGroupProps {
+  options: SelectOptionObject[]
+  title: string | ReactNode
+}
 
-export type SelectOptionProps = ComboboxOptionObject
+export type SelectOptionProps = SelectOptionObject | SelectOptionGroupProps
 
 export interface SelectProps
   extends Omit<ComboboxProps, 'value' | 'defaultValue' | 'onChange'> {
@@ -96,9 +98,71 @@ export interface SelectProps
   onChange?: (value: string) => void
 }
 
-function getOption(value?: string, options?: SelectOptionProps[]) {
-  return value ? { label: getComboboxText(value, options), value } : undefined
+function flattenOptions(options: SelectOptionProps[]) {
+  return options.reduce(
+    (acc: SelectOptionObject[], option: SelectOptionProps) => {
+      const optionAsGroup = option as SelectOptionGroupProps
+      if (optionAsGroup.title) {
+        return [...acc, ...optionAsGroup.options]
+      }
+      return [...acc, option as SelectOptionObject]
+    },
+    []
+  )
 }
+
+function getOption(value?: string, options?: SelectOptionProps[]) {
+  const flattenedOptions = options && flattenOptions(options)
+  return value
+    ? { label: getComboboxText(value, flattenedOptions), value }
+    : undefined
+}
+
+function getFirstOption(options: SelectOptionProps[]): SelectOptionObject {
+  const optionAsGroup = options[0] as SelectOptionGroupProps
+  if (optionAsGroup.title) return optionAsGroup.options[0]
+  return options[0] as SelectOptionObject
+}
+
+const renderOption = (option: SelectOptionObject, index: number) => {
+  if (option.description) {
+    return (
+      <ComboboxOption {...option} key={index} py="xxsmall">
+        <Box>
+          <Heading fontSize="small" fontWeight="semiBold" pb="xxsmall">
+            <ComboboxOptionText />
+          </Heading>
+          <Paragraph variant="subdued" fontSize="small">
+            {option.description}
+          </Paragraph>
+        </Box>
+      </ComboboxOption>
+    )
+  }
+  return <ComboboxOption {...option} key={index} />
+}
+
+const SelectOptionGroupTitle = styled(Heading)`
+  ${comboboxOptionGrid}
+`
+
+SelectOptionGroupTitle.defaultProps = {
+  fontSize: 'xxsmall',
+  fontWeight: 'semiBold',
+  px: 'xsmall',
+  py: 'xxsmall',
+  variant: 'subdued',
+}
+
+const SelectOptionGroup = ({ options, title }: SelectOptionGroupProps) => (
+  <Box py="xxsmall">
+    <SelectOptionGroupTitle>
+      <span />
+      {title}
+    </SelectOptionGroupTitle>
+    {options.map(renderOption)}
+  </Box>
+)
 
 const SelectComponent = forwardRef(
   (
@@ -120,10 +184,12 @@ const SelectComponent = forwardRef(
     ref: Ref<HTMLInputElement>
   ) => {
     const optionValue = getOption(value, options)
-    const defaultOptionValue =
-      getOption(defaultValue, options) || (options && options[0])
+    const nullDefault = (isClearable || placeholder) && !defaultValue
+    const defaultOptionValue = nullDefault
+      ? undefined
+      : getOption(defaultValue, options) || (options && getFirstOption(options))
 
-    function handleChange(option?: ComboboxOptionObject) {
+    function handleChange(option?: SelectOptionObject) {
       const newValue = option ? option.value : ''
       onChange && onChange(newValue)
       onFilter && onFilter('')
@@ -170,9 +236,14 @@ const SelectComponent = forwardRef(
         {!disabled && (
           <ComboboxList persistSelection {...ariaProps}>
             {options && options.length > 0 ? (
-              options.map((option, index: number) => (
-                <ComboboxOption {...option} key={index} />
-              ))
+              options.map((option: SelectOptionProps, index: number) => {
+                const optionAsGroup = option as SelectOptionGroupProps
+                return optionAsGroup.title ? (
+                  <SelectOptionGroup key={index} {...optionAsGroup} />
+                ) : (
+                  renderOption(option as SelectOptionObject, index)
+                )
+              })
             ) : (
               <ListItem fontSize="small" px="medium" py="xxsmall">
                 No options