+++ packages/playground/src/index.tsx
@@ -24,15 +24,13 @@ import { GlobalStyle } from '@looker/components'
 import { theme } from '@looker/design-tokens'
 import { ThemeProvider } from 'styled-components'
 
-import { InputDateDemo } from './InputDate/InputDateDemo'
+import { SelectContent } from './Select/SelectDemo'
 
 const App: React.FC = () => {
   return (
     <ThemeProvider theme={theme}>
-      <>
-        <GlobalStyle />
-        <InputDateDemo />
-      </>
+      <GlobalStyle />
+      <SelectContent />
     </ThemeProvider>
   )
 }